pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "flask>=3.1.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.16",
    "langchain-pinecone>=0.2.6",
    "pinecone>=6.0.2",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
